upgrade to yesod-test-1.5
Integration tests will perform application logging to stdout
This means you can actually figure out why your tests are failing.

A downside is a lot of output interleaved with your tests.
We should address that downside in a future change.
I already have something working locally that I want to try out
for a few weeks and refine a bit.