Add RIO logging to Foundation and runRIOinHandler
It looks to me as if this template will have to be strictly for early adopters
until the deeper conversion of Yesod to RIO has been completed. All the
differences are superficial but it looks like there are more than enough of
them to confuse without a special guide for the unwary.

The major issue is that the handler is just not a RIO so application developers
are going to have to apply converters to their RIO based handlers to insert
them into Yesod 1.6.

The converter I have constructed here is based on the one I have written for
my own application is based merely on the `App` foundation, but it should
probably be based on `HandlerData`, to give us the (projected) Yesod 1.7
handlers. (I am happy to do this in the next patch if @snoyberg agrees --
I would expect it to be a straightforward generalisation.)

The other immediate issue is that, AFAICS, the RIO loggers and monad-logger
loggers are incompatible, so the Yesod porcelain will continue to use the latter
while RIO handlers would use the former. (Which works for me, but might come
across as clunky for others.)