Warn that TRUNCATEing many tables with Postgres can become slow
I think the vast, vast majority of yesod-scaffold users will never grow to the size where they need to care about this. But for the handful that do, it should save them a lot of time.

I could have made a wiki page about this and cleaned it all up, but I think it's not really worth it for how much of an edge case this is. To reach this point, you need to be a pretty advanced Yesod user